Update mysql provisioner

+++ driver/mysql/mysql_driver.go
@@ -6,7 +6,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"strings"
 
 	"github.com/hpcloud/cf-usb/driver/mysql/mysqlprovisioner"
 	"github.com/hpcloud/cf-usb/lib/config"
@@ -43,72 +42,95 @@ func NewMysqlDriver(logger lager.Logger) driver.Driver {
 func (e *MysqlDriver) Init(configuration config.DriverProperties, response *string) error {
 	err := json.Unmarshal(*configuration.DriverConfiguration, &e)
 	e.logger.Info("Mysql Driver initializing")
-	e.db = mysqlprovisioner.New(e.User, e.Pass, e.Host+":"+e.Port)
+	e.db, err = mysqlprovisioner.New(e.User, e.Pass, e.Host+":"+e.Port)
 	return err
 }
 
 func (e *MysqlDriver) Provision(request model.DriverProvisionRequest, response *string) error {
+	created, err := e.db.IsDatabaseCreated(request.InstanceID)
 
-	err := e.db.CreateDatabase(request.InstanceID)
 	if err != nil {
-		if strings.Contains(err.Error(), "database exists") {
-			return brokerapi.ErrInstanceAlreadyExists
-		}
-		return brokerapi.ErrInstanceAlreadyExists
+		return err
 	}
 
-	*response = fmt.Sprintf("http://localhost/instance/%s/service/%s/dashboard", request.InstanceID, request.ServiceDetails.ID)
+	if created {
+		return brokerapi.ErrInstanceAlreadyExists
+	} else {
+		err := e.db.CreateDatabase(request.InstanceID)
+		if err != nil {
+			return err
+		}
+	}
 
 	return nil
 }
 
 func (e *MysqlDriver) Deprovision(request model.DriverDeprovisionRequest, response *string) error {
+	created, err := e.db.IsDatabaseCreated(request.InstanceID)
 
-	err := e.db.DeleteDatabase(request.InstanceID)
 	if err != nil {
-		if strings.Contains(err.Error(), "database doesn't exist") {
-			return brokerapi.ErrInstanceDoesNotExist
-		}
 		return err
 	}
-
-	*response = fmt.Sprintf("Deprovisioned %s", request.InstanceID)
-
+	if created {
+		err := e.db.DeleteDatabase(request.InstanceID)
+		if err != nil {
+			return err
+		}
+		*response = fmt.Sprintf("Deprovisioned %s", request.InstanceID)
+	} else {
+		return brokerapi.ErrInstanceDoesNotExist
+	}
 	return nil
 }
 
 func (e *MysqlDriver) Bind(request model.DriverBindRequest, response *json.RawMessage) error {
+
 	username := request.InstanceID + "-" + request.BindingID
 	password := e.secureRandomString(32)
 
-	err := e.db.CreateUser(request.InstanceID, username, password)
-	if err != nil {
-		return brokerapi.ErrBindingAlreadyExists
-	}
-
-	data := MysqlBindingCredentials{
-		Host:             e.Host,
-		Port:             e.Port,
-		Username:         username,
-		Password:         password,
-		ConnectionString: generateConnectionString(e.Host, e.Port, request.InstanceID, username, password),
-	}
-	marhsaled, err := json.Marshal(data)
+	created, err := e.db.IsUserCreated(username)
 	if err != nil {
 		return err
 	}
-	response = (*json.RawMessage)(&marhsaled)
 
+	if created {
+		return brokerapi.ErrBindingAlreadyExists
+	} else {
+
+		err := e.db.CreateUser(request.InstanceID, username, password)
+		if err != nil {
+			return err
+		}
+		data := MysqlBindingCredentials{
+			Host:             e.Host,
+			Port:             e.Port,
+			Username:         username,
+			Password:         password,
+			ConnectionString: generateConnectionString(e.Host, e.Port, request.InstanceID, username, password),
+		}
+		marhsaled, err := json.Marshal(data)
+		if err != nil {
+			return err
+		}
+		response = (*json.RawMessage)(&marhsaled)
+	}
 	return nil
 }
 
 func (e *MysqlDriver) Unbind(request model.DriverUnbindRequest, response *string) error {
 	username := request.InstanceID + "-" + request.BindingID
-
-	err := e.db.DeleteUser(username)
+	created, err := e.db.IsUserCreated(username)
 	if err != nil {
+		return err
+	}
+	if created {
+		err := e.db.DeleteUser(username)
+		if err != nil {
+			return err
+		}
+	} else {
 		return brokerapi.ErrBindingDoesNotExist
 	}
-
+	*response = fmt.Sprintf("Unbound %s", username)
 	return nil
 }
